Move top bar model surface contracts
