+++ docs/DECISIONS.md
@@ -93,6 +93,7 @@
 - Runtime issue retry is a command, not a view-only callback. The Retry action prepares the latest non-empty user turn from the selected transcript and submits it through the normal composer path, so slash commands, tool dispatch, safety review, persistence, and future telemetry stay identical to a user pressing Send again.
 - Runtime issue recovery actions should take the user directly to the corrective control. Malformed model/tool responses and provider rate limits use `Switch model` to open the same searchable top-bar model picker and focus model search, instead of introducing a second recovery modal or leaving the user to find the picker manually.
 - Model picker presentation is built by `WorkspaceModelCatalogSurfaceBuilder`. It owns label/category construction, selected/default badge comparison, favorite/recent deduping, current-model fallback insertion, and row badge construction; `WorkspaceSurface.swift` only supplies catalog/config/thread-history inputs and consumes the finished label/categories.
+- Top-bar and model-picker surface contracts live in `QuillCodeTopBarSurface.swift`. `TopBarSurface`, model category rows, model metadata rows, option compatibility decoding, option metadata copy, and searchable model filtering stay beside the top-bar/model-picker UI boundary; `WorkspaceSurface.swift` should only carry the aggregate `topBar` payload and assemble it through the focused catalog builder.
 - Command palette presentation is built by `WorkspaceCommandSurfaceBuilder`. `WorkspaceSurface.swift` supplies selected thread/project/sidebar/runtime facts, `WorkspaceCommandSurfaceBuilder` owns command rows, categories, availability, and search keywords, and `WorkspaceCommandPlan` remains the separate action reducer that maps command IDs to model mutations and tool dispatch.
 - Review diff presentation is built by `WorkspaceReviewSurfaceBuilder`. `WorkspaceSurface.swift` supplies tool cards and thread events, while the builder owns latest successful `host.git.diff` selection, diff parsing, file/line review comment attachment, timestamp ordering, and line-kind filtering.
 - Context pressure presentation is built by `WorkspaceContextBannerBuilder`. `WorkspaceSurface.swift` supplies the selected thread, while the builder owns empty-thread hiding, context estimate calculation, warning thresholds, full-context copy, and New/Fork/Compact command surfaces.
